+++ ImageProcessor/src/image/images_in_memory/pgm/InMemoryPgmBinary.java
@@ -3,7 +3,7 @@ package image.images_in_memory.pgm;
 import image.signatures.FormatType;
 
 public class InMemoryPgmBinary extends InMemoryPgm {
-    public InMemoryPgmBinary(int width, int height, int maxValue) {
+    public InMemoryPgmBinary(int width, int height, short maxValue) {
         super(width, height, maxValue);
     }
